+++ exploded/Assets/PlayServicesResolver/Editor/ResolverVer1_1.cs
@@ -21,10 +21,27 @@ namespace GooglePlayServices
     using System.Collections.Generic;
     using Google.JarResolver;
     using System.IO;
+    using System.Xml;
 
     [InitializeOnLoad]
     public class ResolverVer1_1 : DefaultResolver
     {
+        // Caches data associated with an aar so that it doesn't need to be queried to determine
+        // whether it should be expanded / exploded if it hasn't changed.
+        private class AarExplodeData
+        {
+            // Time the file was modified the last time it was inspected.
+            public System.DateTime modificationTime;
+            // Whether the AAR file should be expanded / exploded.
+            public bool explode;
+        }
+
+        private Dictionary<string, AarExplodeData> aarExplodeData =
+            new Dictionary<string, AarExplodeData>();
+        // File used to to serialize aarExplodeData.  This is required as Unity will reload classes
+        // in the editor when C# files are modified.
+        private string aarExplodeDataFile = Path.Combine("ProjectSettings",
+                                                         "GoogleAarExplodeCache.xml");
 
         private const int MajorVersion = 1;
         private const int MinorVersion = 1;
@@ -32,7 +49,89 @@ namespace GooglePlayServices
 
         static ResolverVer1_1()
         {
-            PlayServicesResolver.RegisterResolver(new ResolverVer1_1());
+            ResolverVer1_1 resolver = new ResolverVer1_1();
+            resolver.LoadAarExplodeCache();
+            PlayServicesResolver.RegisterResolver(resolver);
+        }
+
+        /// <summary>
+        /// Load data cached in aarExplodeDataFile into aarExplodeData.
+        /// </summary>
+        private void LoadAarExplodeCache()
+        {
+            if (!File.Exists(aarExplodeDataFile)) return;
+
+            XmlTextReader reader = new XmlTextReader(new StreamReader(aarExplodeDataFile));
+            while (reader.Read())
+            {
+                if (reader.NodeType == XmlNodeType.Element && reader.Name == "aars")
+                {
+                    while (reader.Read())
+                    {
+                        if (reader.NodeType == XmlNodeType.Element && reader.Name == "explodeData")
+                        {
+                            string aar = "";
+                            AarExplodeData aarData = new AarExplodeData();
+                            do
+                            {
+                                if (!reader.Read()) break;
+                                if (reader.NodeType == XmlNodeType.Element)
+                                {
+                                    string elementName = reader.Name;
+                                    if (reader.Read() && reader.NodeType == XmlNodeType.Text)
+                                    {
+                                        if (elementName == "aar")
+                                        {
+                                            aar = reader.ReadContentAsString();
+                                        }
+                                        else if (elementName == "modificationTime")
+                                        {
+                                            aarData.modificationTime =
+                                                reader.ReadContentAsDateTime();
+                                        }
+                                        else if (elementName == "explode")
+                                        {
+                                            aarData.explode = reader.ReadContentAsBoolean();
+                                        }
+                                    }
+                                }
+                            } while (!(reader.Name == "explodeData" &&
+                                       reader.NodeType == XmlNodeType.EndElement));
+                            if (aar != "") aarExplodeData[aar] = aarData;
+                        }
+                    }
+                }
+            }
+        }
+
+        /// <summary>
+        /// Save data from aarExplodeData into aarExplodeDataFile.
+        /// </summary>
+        private void SaveAarExplodeCache()
+        {
+            if (File.Exists(aarExplodeDataFile))
+            {
+                File.Delete(aarExplodeDataFile);
+            }
+            XmlTextWriter writer = new XmlTextWriter(new StreamWriter(aarExplodeDataFile));
+            writer.WriteStartElement("aars");
+            foreach (KeyValuePair<string, AarExplodeData> kv in aarExplodeData)
+            {
+                writer.WriteStartElement("explodeData");
+                writer.WriteStartElement("aar");
+                writer.WriteValue(kv.Key);
+                writer.WriteEndElement();
+                writer.WriteStartElement("modificationTime");
+                writer.WriteValue(kv.Value.modificationTime);
+                writer.WriteEndElement();
+                writer.WriteStartElement("explode");
+                writer.WriteValue(kv.Value.explode);
+                writer.WriteEndElement();
+                writer.WriteEndElement();
+            }
+            writer.WriteEndElement();
+            writer.Flush();
+            writer.Close();
         }
 
         #region IResolver implementation
@@ -69,6 +168,8 @@ namespace GooglePlayServices
             // e.g. ${applicationId}.  Unity does not understand how to process
             // these, so we handle it here.
             ProcessAars(destinationDirectory);
+
+            SaveAarExplodeCache();
         }
 
         #endregion
@@ -109,14 +210,61 @@ namespace GooglePlayServices
         }
 
         /// <summary>
-        /// Shoulds the explode the aar file.
+        /// Determined whether an aar file should be exploded (extracted).
+        ///
+        /// This is required for some aars so that the Unity Jar Resolver can perform variable
+        /// expansion on manifests in the package before they're merged by aapt.
         /// </summary>
-        /// <returns><c>true</c>, if explode was shoulded, <c>false</c> otherwise.</returns>
+        /// <returns><c>true</c>, if the aar should be exploded, <c>false</c> otherwise.</returns>
         /// <param name="aarFile">The aar file.</param>
         internal virtual bool ShouldExplode(string aarFile)
         {
-            return aarFile.Contains("play-services-measurement") ||
-            !SupportsAarFiles;
+            AarExplodeData aarData = new AarExplodeData();
+            aarData.explode = !SupportsAarFiles;
+            if (!aarData.explode)
+            {
+                AarExplodeData retrievedAarData = null;
+                if (aarExplodeData.TryGetValue(aarFile, out retrievedAarData))
+                {
+                    System.DateTime modificationTime = File.GetLastWriteTime(aarFile);
+                    if (modificationTime.CompareTo(aarData.modificationTime) <= 0)
+                    {
+                        aarData = retrievedAarData;
+                    }
+                }
+            }
+            if (!aarData.explode)
+            {
+                string temporaryDirectory = CreateTemporaryDirectory();
+                if (temporaryDirectory == null) return false;
+                string manifestFilename = "AndroidManifest.xml";
+                try
+                {
+                    if (ExtractAar(aarFile, new string[] {manifestFilename},
+                                   temporaryDirectory))
+                    {
+                        string manifestPath = Path.Combine(temporaryDirectory,
+                                                           manifestFilename);
+                        if (File.Exists(manifestPath))
+                        {
+                            string manifest = File.ReadAllText(manifestPath);
+                            aarData.explode = manifest.IndexOf("${applicationId}") >= 0;
+                        }
+                        aarData.modificationTime = File.GetLastWriteTime(aarFile);
+                    }
+                }
+                catch (System.Exception e)
+                {
+                    UnityEngine.Debug.Log("Unable to examine AAR file " + aarFile + ", err: " + e);
+                    throw e;
+                }
+                finally
+                {
+                    DefaultResolver.DeleteFully(temporaryDirectory);
+                }
+            }
+            aarExplodeData[aarFile] = aarData;
+            return aarData.explode;
         }
 
         /// <summary>